Implement as test rendering pipeline.
Implemented rendering pipeline and added unittests.  Images are compared using root mean square to make sure they are approximately the same (as the PIL library save with a color profile which slightly lightens the image).

Fixed Integer typecasting in matrix_manager for older versions of python.

+++ tile_manager/matrix_manager.py
@@ -122,8 +122,8 @@ class MatrixInterface(object):
     if self.led_rows == 8:
       self.width = self.height = self.led_rows
     
-    tiles_x = self.width / self.tile_size
-    tiles_y = self.height / self.tile_size
+    tiles_x = int(self.width / self.tile_size)
+    tiles_y = int(self.height / self.tile_size)
     self.shape = [[None for i in range(tiles_x)] for i in range(tiles_y)]
 
   def FillScreen(self, fill=(0,0,0)):
